+++ src/lib.rs
@@ -62,7 +62,7 @@ pub use crate::{
     error::Error,
     event::Event,
     metadata::Value as MetadataValue,
-    transaction::{Flag, Posting, Transaction},
+    transaction::{Flag, Posting, PostingPrice, Transaction},
 };
 use nom::{
     branch::alt,